requirements.txt
APScheduler==3.9.1.post1
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==3.2
django-admin-honeypot==1.1.0
django-apscheduler==0.6.2
django-axes==5.40.1
django-db-logger==0.1.12
django-dbbackup==4.0.2
django-ipware==4.0.2
django-jazzmin==2.6.0
django-tinymce==3.5.0
gunicorn==20.1.0
idna==3.4
psycopg2-binary==2.9.5
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stripe==5.0.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
